feat: added listOf, setOf and arrayOf functions + refactor: move print & shuffleArray functions to util package
