+++ core/src/main/java/com/instancify/scriptify/core/script/function/impl/util/ScriptFunctionPrint.java
@@ -1,4 +1,4 @@
-package com.instancify.scriptify.core.script.function.impl;
+package com.instancify.scriptify.core.script.function.impl.util;
 
 import com.instancify.scriptify.api.script.Script;
 import com.instancify.scriptify.api.script.function.ScriptFunction;
